% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartanodown.R
\docType{package}
\name{spartanodown}
\alias{spartanodown}
\alias{spartanodown-package}
\title{spartanodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
spartanodown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
